These changes incorperate a fairly complete c-interface layer, although maybe more restructuring should be done of this package.
git-svn-id: https://moleculizer.svn.sourceforge.net/svnroot/moleculizer/branches/libmoleculizer@529 e2b6e7e4-f0f2-0310-8dec-eb45598a66c7

+++ src/cpx/isoSearchImpl.hh
@@ -386,7 +386,7 @@ namespace cpx
 
         std::vector<int>::const_iterator begin_first( labelingMap1.begin() ), begin_second( labelingMap2.begin() );
 
-        for ( unsigned int partNdx = 0; partNdx != partitionMap.size(); ++partNdx )
+        for ( int partNdx = 0; partNdx != (int) partitionMap.size(); ++partNdx )
         {
             if ( partitionMap[partNdx] == 0 ) continue;
